feat!: refactor the input processor
This is a complete redesign of the input scanner, using logic derived
from tcell.  It eliminates the per-terminal special parsing for keys,
so that we can process keys from almost anything.

This adds parsing support (not enabled yet) for CSI-u (kitty protocol),
Xterm modifierOtherKeys, and win32-input-mode protocols.  (These will
be enabled in follow up.)

As a bonus, this adds suppot for focus reporting, so that we get to
now when the application receives or loses focus.  The mouse demo is
updated to reflect that.

One consequence of this is that the key definitions for all the various
control keys are gone, as these are now conveyed as a modified Key rune.

The key definitions (and also PadChar) are removed from the terminfo
definitions.

This is a breaking change, but we are still pre-1.0 so we didn't have any
promises anyway.