feat!: Major refactoring in progress for the terminfo subsystem
We're moving to delete the terminfo database, and this is more work
on that path.  We have better support for standard escapes, as the
terminals have common and compatible functionality (the terminfo DB
had pointless, from our perspective, differences).

This also eliminates the automargin bug hackery, as no supported
terminals actually need it.

The terminfo capabilities have shrunk considerably.  We still need
to sort out a few last bits, before delete this entirely.

This also has some termio fixes.