fix: partial writes to terminal can crash
The TTY driver does not guarantee a complete write, and the rawWrite
API (which we used over concerns about data transformation) throws if
the data is not fully written.  The "write" API is free from this flaw,
and if we only use strings, then the data is passed through unadulterated.