feat: Completely refactor termio processing and eliminate threads (fixes #8)
This now uses a new api for obtaining events, which allows the application
to run in a single event loop without any locks or threading.  Once small
consequence of this is that if the application wants to work in an MT way
with the screen, then it will need to provide its own guards.

This design means that this can now be used in designs that may not be conducive
to using runtime threads (such as reactor designs found in Weka's mecca library),
and it also should make debugging much simpler.