feat!: add convenience methods for writing to screen (fixes #3 fixes #4)
This adds a nice string-aware, grapheme-aware mechansim for writing to the
screen.  A stateful cursor and a stateful style are used, which then allows
write to be used simply.

The setStyle method is replaced by just a normal style property.

A range API is also provided, but only works for simple cases that do not use
grapheme clusters (however most interfaces don't need grapheme clusters, and
indeed many terminal emulators don't treat them properly anyway.)

While here, a number of methods were marked `@safe`, although this is not
comprehensive.