feat: expose file based constructor
This should allow applications to integrate dcell into external
polling loops.  These applications should use the events and
waitForEvent APIs to adjust their poll intervals as needed.

A cautionary note is given for macOS, because macOS bizarrely does
not support poll or kqueue with tty devices.  (The only way to
work on such systems is to use a separate thread and the self-pipe
trick.)